TEXTURES GOOD SUR LES MURS

+++ include/struct_cub3D.h
@@ -131,13 +131,13 @@ typedef struct		s_sprite
 
 typedef struct		s_texture
 {
-	float			wallx; // valeur où le mur a été touché : coordonnée y si side == 0, coordonnée x si side == 1
+	double			wallx; // valeur où le mur a été touché : coordonnée y si side == 0, coordonnée x si side == 1
 	int			text_dir; // direction NO, S, EA, WE de la texture
 	int			x; // coordonnee x de la texture
 	int			y; // coordonnee y de la texture
-	int			step; // de combien augmenter les coordonnées de la texture pour chaque pixel
-	int			position; // coordonnee de depart
-	float		shade;
+	double			step; // de combien augmenter les coordonnées de la texture pour chaque pixel
+	double		position; // coordonnee de depart
+	double		shade;
 	int			*color;
 }					t_texture;
